+++ CHANGES.rst
@@ -32,6 +32,8 @@ Unreleased
 - Remove ``Comment.gild``, ``Redditor.gild``, and ``Submission.gild`` methods.
 - Remove ``Redditor.gilded`` and ``Subreddit.gilded`` methods.
 - Remove ``Redditor.gildings`` method.
+- Remove ``Subreddit.mod.inbox``, ``Subreddit.mod.unread``, and
+  ``Subreddit.mod.stream.unread`` methods.
 - Remove ``Subreddits.gold`` method.
 - Remove :class:`.Reddit` keyword argument ``token_manager`` and all associated token
   managers.